changed all C++ ints to int64_t for hhvm extension

+++ hhvm/extension/newrelic.cpp
@@ -98,27 +98,27 @@ static int64_t HHVM_FUNCTION(hhvm_newrelic_transaction_begin) {
 	return transaction_id;
 }
 
-static int HHVM_FUNCTION(hhvm_newrelic_transaction_notice_error, const String& exception_type, const String& error_message, const String& stack_trace, const String& stack_frame_delimiter) {
+static int64_t HHVM_FUNCTION(hhvm_newrelic_transaction_notice_error, const String& exception_type, const String& error_message, const String& stack_trace, const String& stack_frame_delimiter) {
 	return newrelic_transaction_notice_error(NEWRELIC_AUTOSCOPE, exception_type.c_str(), error_message.c_str(), stack_trace.c_str(), stack_frame_delimiter.c_str());
 }
 
-static int HHVM_FUNCTION(hhvm_newrelic_transaction_add_attribute, const String& name, const String& value) {
+static int64_t HHVM_FUNCTION(hhvm_newrelic_transaction_add_attribute, const String& name, const String& value) {
 	return newrelic_transaction_add_attribute(NEWRELIC_AUTOSCOPE, name.c_str(), value.c_str());
 }
 
-static int HHVM_FUNCTION(hhvm_newrelic_transaction_set_name, const String& name) {
+static int64_t HHVM_FUNCTION(hhvm_newrelic_transaction_set_name, const String& name) {
 	return newrelic_transaction_set_name(NEWRELIC_AUTOSCOPE, name.c_str());
 }
 
-static int HHVM_FUNCTION(hhvm_newrelic_transaction_set_request_url, const String& request_url) {
+static int64_t HHVM_FUNCTION(hhvm_newrelic_transaction_set_request_url, const String& request_url) {
 	return newrelic_transaction_set_request_url(NEWRELIC_AUTOSCOPE, request_url.c_str());
 }
 
-static int HHVM_FUNCTION(hhvm_newrelic_transaction_set_max_trace_segments, int max_trace_segments) {
+static int64_t HHVM_FUNCTION(hhvm_newrelic_transaction_set_max_trace_segments, int64_t max_trace_segments) {
 	return newrelic_transaction_set_max_trace_segments(NEWRELIC_AUTOSCOPE, max_trace_segments);
 }
 
-static int HHVM_FUNCTION(hhvm_newrelic_transaction_end) {
+static int64_t HHVM_FUNCTION(hhvm_newrelic_transaction_end) {
 	return newrelic_transaction_end(NEWRELIC_AUTOSCOPE);
 }
 
@@ -130,7 +130,7 @@ static int64_t HHVM_FUNCTION(hhvm_newrelic_segment_datastore_begin, const String
 	return newrelic_segment_datastore_begin(NEWRELIC_AUTOSCOPE, NEWRELIC_AUTOSCOPE, table.c_str(), operation.c_str());
 }
 
-static int HHVM_FUNCTION(hhvm_newrelic_segment_end, int64_t id) {
+static int64_t HHVM_FUNCTION(hhvm_newrelic_segment_end, int64_t id) {
 	return newrelic_segment_end(NEWRELIC_AUTOSCOPE, id);
 }
